Move FlatShippingRatesInputCards into ShippingRateSection component folder.
