--- js/src/components/flat-shipping-rates-input-cards/index.js
@@ -1,1 +1,0 @@
-export { default } from './flat-shipping-rates-input-cards';
