+++ js/src/components/shipping-rate-section/minimum-order-card/minimum-order-input-control.scss
